feat(base-app): templates page to Tailwind WIP
