+++ apps/base-app/src/components/TemplateSelector.vue
@@ -1,7 +1,7 @@
 <script lang="ts">
 import { defineComponent, type PropType } from 'vue'
 import { BCol, BRow } from 'bootstrap-vue'
-import TemplateSelectorOption from './TemplateSelectorOption.vue'
+import TemplateSelectorOption from '../entities/textSettingsTemplate/ui/TextSettingsTemplatePreview.vue'
 import type { SettingsTemplate } from '@/interfaces/settingsTemplate'
 
 const TemplateSelector = defineComponent({
