+++ apps/base-app/src/entities/textSettingsTemplate/index.ts
@@ -1,1 +1,2 @@
 export type { TextSettingsTemplate } from './model/textSettingsTemplate'
+export { default as TextSettingsTemplatePreview } from './ui/TextSettingsTemplatePreview.vue'
