+++ apps/base-app/src/features/textSettingsAdaptationPreview/index.ts
@@ -1,4 +1,1 @@
-// TODO: maybe rename the module to textSettingsAdaptationPreview?
 export { default as TextSettingsAdaptationPreview } from './ui/TextSettingsAdaptationPreview.vue'
-// TODO: encapsulate the usage of the function inside the module
-export { adaptHtmlElementAsyncFn } from './lib/visualEngine/adaptHtmlElementAsync'
