+++ apps/base-app/src/shared/lib/textAdaptation.ts
